pyproject.toml
[project]
name = "game-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.17.0",
    "asyncpg>=0.30.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "sqlalchemy[asyncio]>=2.0.38",
]
